+++ packages/langium/src/index.ts
@@ -4,30 +4,31 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-export * from './dependency-injection';
 export * from './default-module';
+export * from './dependency-injection';
 export * from './documents/document';
 export * from './documents/document-builder';
-export * from './services';
-export * from './syntax-tree';
 export * from './generator/node';
 export * from './generator/node-processor';
 export * from './generator/template-string';
-export * from './grammar/langium-grammar-module';
 export * from './grammar/generated/ast';
 export * from './grammar/generated/grammar-access';
 export * from './grammar/generated/module';
 export * from './grammar/generated/parser';
 export * from './grammar/grammar-access';
 export * from './grammar/grammar-util';
-export * from './service/language-server';
-export * from './service/validation/document-validator';
-export * from './service/validation/validation-registry';
+export * from './grammar/langium-grammar-module';
+export * from './lsp/completion/completion-provider';
+export * from './lsp/completion/follow-element-computation';
+export * from './lsp/completion/rule-interpreter';
+export * from './lsp/language-server';
+export * from './lsp/validation/document-validator';
+export * from './lsp/validation/validation-registry';
 export * from './parser/langium-parser';
-export * from './service/completion/completion-provider';
-export * from './service/completion/follow-element-computation';
-export * from './service/completion/rule-interpreter';
+export * from './serializer/json-serializer';
+export * from './services';
+export * from './syntax-tree';
 export * from './utils/ast-util';
 export * from './utils/cst-util';
 export * from './utils/stream';
-export * from './serializer/json-serializer';
+
