+++ packages/langium/src/lsp/goto.ts
@@ -6,11 +6,11 @@
 
 import { Location, Range, TextDocumentPositionParams } from 'vscode-languageserver';
 import { LangiumDocument } from '../documents/document';
+import { NameProvider } from '../references/naming';
+import { References } from '../references/references';
 import { LangiumServices } from '../services';
 import { CstNode } from '../syntax-tree';
 import { findLeafNodeAtOffset } from '../utils/ast-util';
-import { NameProvider } from './naming';
-import { References } from './references';
 
 export interface GoToResolver {
     goToDeclaration(document: LangiumDocument, position: TextDocumentPositionParams): Location[]
